requirements.txt
slack-sdk>=3.37.0
slack-bolt>=1.26.0
python-dotenv>=1.0.0
